build(Docker): add a user with the same host's user ID
Update the compose file to pass args to the Dockerfile.

Update the compose file to set the host's user as the default user on the container.

Update the compose file to add an environment file.

Update the Dockerfile to create a user with the same host's user ID and add this user to the group 'users'.