feat(Emacs): Create the Emacs config files
Create init.el file.

Create a config.org file.

Add the Org-Babel in the init.el file to use the code blocks of the config.org file to overwrite the init.el file and build the Emacs
config when the config.org file was saved.

Bind the directory /config/emacs between the host and the container.